refactor: change folder for buttons, headers, popup, avatar components; refactor: response for logout
